requirements.txt
asgiref==3.7.2
distlib==0.3.6
Django==4.2.2
django-mail-templated==2.6.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
environs==9.5.0
filelock==3.12.0
inflection==0.5.1
marshmallow==3.19.0
packaging==23.1
platformdirs==3.5.1
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
virtualenv==20.23.0
